pyproject.toml
[tool.poetry]
name = "codebox"
version = "0.1.0"
description = ""
authors = ["usakaonick@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
jupyterlab = "^4.0.9"
aiohttp = "^3.9.1"
codeboxapi = "^0.0.19"
websockets = "11.0.3"
jupyter-kernel-gateway = "^2.5.2"
numpy = "^1.26.2"
pyaudio = "^0.2.14"
speechrecognition = "^3.10.1"
google-cloud-speech = "^2.23.0"
redis = "^5.0.1"
openai = "1.6.1"
requests = "^2.31.0"
proxychains = "^1.0"
streamlit = "^1.30.0"
plotly = "^5.18.0"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
sidecar = "^0.7.0"
ipympl = "^0.9.3"
jupyterlab-lsp = "^5.0.2"
litellm = "^1.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
